+++ test/fixtures/queries/select_star.sql
@@ -1,0 +1,3 @@
+SELECT *
+FROM test/fixtures/tables/people
+WHERE age > 30