+++ ChangeLog.md
@@ -4,6 +4,7 @@
 * SELECT Expressions
 * SELECT AS
 * SELECT Expression type checking
+* SELECT * support
 
 ### v0.0.6
 * Type Checking of WHERE clause